+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 - Fixed a bug where Neo fields which were set to translatable on a Craft 2 install would not have their propagation method properly set when upgrading to Craft 3
+- Fixed a bug when editing a Neo field's settings, where the Field Layout tab would display and be selectable if a block type group was selected
 - Some minor changes to JavaScript code related to new block button enabling/disabling, block dragging, and Neo-to-Matrix conversion
 
 ## 2.8.4 - 2020-08-12